+++ application/src/test/java/org/thingsboard/server/controller/AbstractControllerTest.java
@@ -16,16 +16,25 @@
 package org.thingsboard.server.controller;
 
 import lombok.extern.slf4j.Slf4j;
+import org.junit.After;
+import org.junit.Before;
 import org.junit.runner.RunWith;
 import org.springframework.boot.test.context.SpringBootContextLoader;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.test.context.web.WebAppConfiguration;
+import org.springframework.web.socket.config.annotation.EnableWebSocket;
+
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.util.concurrent.TimeUnit;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 @ActiveProfiles("test")
 @RunWith(SpringRunner.class)
@@ -33,9 +42,49 @@ import org.springframework.test.context.web.WebAppConfiguration;
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_CLASS)
 @Configuration
 @ComponentScan({"org.thingsboard.server"})
-@WebAppConfiguration
-@SpringBootTest()
+@EnableWebSocket
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Slf4j
-public abstract class AbstractControllerTest extends AbstractWebsocketTest  {
+public abstract class AbstractControllerTest extends AbstractWebTest {
+
+    public static final String WS_URL = "ws://localhost:";
+
+    @LocalServerPort
+    protected int wsPort;
+
+    private TbTestWebSocketClient wsClient; // lazy
+
+    public TbTestWebSocketClient getWsClient() {
+        if (wsClient == null) {
+            synchronized (this) {
+                try {
+                    if (wsClient == null) {
+                        wsClient = buildAndConnectWebSocketClient();
+                    }
+                } catch (Exception e) {
+                    throw new RuntimeException(e);
+                }
+            }
+        }
+        return wsClient;
+    }
+
+    @Before
+    public void beforeWsTest() throws Exception {
+        // placeholder
+    }
+
+    @After
+    public void afterWsTest() throws Exception {
+        if (wsClient != null) {
+            wsClient.close();
+        }
+    }
+
+    private TbTestWebSocketClient buildAndConnectWebSocketClient() throws URISyntaxException, InterruptedException {
+        TbTestWebSocketClient wsClient = new TbTestWebSocketClient(new URI(WS_URL + wsPort + "/api/ws/plugins/telemetry?token=" + token));
+        assertThat(wsClient.connectBlocking(TIMEOUT, TimeUnit.SECONDS)).isTrue();
+        return wsClient;
+    }
 
 }
